melhorando flash e layout de alerta, aqui já esta no ar com calendario e tudo mais!

+++ src/controllers/ClientController.php
@@ -16,12 +16,18 @@ class ClientController extends Controller {
         }
     }
     public function viewClient(){
+        $flash  ='';
+        if(!empty($_SESSION['flash'])){
+            $flash = $_SESSION['flash'];
+            $_SESSION['flash'] = '';
+        }
         $id_user = $this->loggedUser->id;
         $name_user = $this->loggedUser->name;
         $clients = ClientHandler::getAllClient();
         $this->render('viewclient', [
             'loggedUser' => $this->loggedUser,
-            'client' => $clients
+            'client' => $clients,
+            'flash' => $flash
         ]);
     }
     public function toreceive(){
@@ -120,85 +126,109 @@ class ClientController extends Controller {
         $phone = filter_input(INPUT_POST, 'phone');
         $service = filter_input(INPUT_POST, 'service');
         $comment = filter_input(INPUT_POST, 'comment');
-        if(!empty($service)){
+        if(!empty($service) && $service != $client->service){
             ClientHandler::editService($service, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
-        if(!empty($phone)){
+        if(!empty($phone) && $phone != $client->phone){
             ClientHandler::editPhone($phone, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
-        if(!empty($email)){
+        if(!empty($email) && $email != $client->email){
             ClientHandler::editEmail($email, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
-        if(!empty($comment)){
+        if(!empty($comment) && $comment != $client->comment){
             ClientHandler::editComment($comment, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
         if(isset($_FILES['rg']) && !empty($_FILES['rg']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/rg/'.$client->rg);
+            if(!empty($client->rg)){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/rg/'.$client->rg);
             $newRg = $_FILES['rg'];
+            }
             /// DESENVOLVIMENTO 
             $rgName = ClientHandler::ImageNoCut($newRg, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\rg');
             
             $rg = $rgName;
             ClientHandler::editRg($rg,$id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
         if(isset($_FILES['cpf']) && !empty($_FILES['cpf']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/cpf/'.$client->cpf);
+            if(!empty($client->cpf)){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/cpf/'.$client->cpf);
+            }
             $newCpf = $_FILES['cpf'];
             /// DESENVOLVIMENTO 
             $cpfName = ClientHandler::ImageNoCut($newCpf, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\cpf');
             
             $cpf = $cpfName;
             ClientHandler::editCpf($cpf, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
         if(isset($_FILES['photo_client']) && !empty($_FILES['photo_client']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/self/'.$client->photo_client);
+            if(!empty($client->photo_client)){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/self/'.$client->photo_client);
+            }
             $newPhoto_client = $_FILES['photo_client'];
             /// DESENVOLVIMENTO 
             $photo_clientName = ClientHandler::ImageNoCut($newPhoto_client, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\self');
             
             $photo_client = $photo_clientName;
             ClientHandler::editSelf($photo_client, $id);
+            $_SESSION['flash'] = 'Atualizado com sucesso !';
         }
         if(isset($_FILES['extract']) && !empty($_FILES['extract']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/extrato/'.$client->extract);
+            if(!empty($client->extract)){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/extrato/'.$client->extract);
+            }
             $newExtract = $_FILES['extract'];
             /// DESENVOLVIMENTO 
             $extractName = ClientHandler::ImageNoCut($newExtract, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\extrato');
             
             $extract = $extractName;
             ClientHandler::editExtract($extract, $id);
-            $_SESSION['flash'] = 'Extrato atualizado com sucesso';
+            $_SESSION['flash'] = 'Atualizado com sucesso';
         }
         if(isset($_FILES['residence']) && !empty($_FILES['residence']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/comprovante/'.$client->residence);
+            if($client->residence){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/comprovante/'.$client->residence);
+            }
             $newResidence = $_FILES['residence'];
             /// DESENVOLVIMENTO 
             $residenceName = ClientHandler::ImageNoCut($newResidence, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\comprovante');
             
             $residence = $residenceName;
             ClientHandler::editResidence($residence, $id);
+            $_SESSION['flash'] = 'Atualizada com sucesso';
         }
         if(isset($_FILES['mirror']) && !empty($_FILES['mirror']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/zap/'.$client->mirror);
+            if($client->mirror){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/zap/'.$client->mirror);
+            }
             $newMirror = $_FILES['mirror'];
             /// DESENVOLVIMENTO 
             $mirrorName = ClientHandler::ImageNoCut($newMirror, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\espelho');
             
             $mirror = $mirrorName;
             ClientHandler::editMirror($mirror, $id);
+            $_SESSION['flash'] = 'Atualizada com sucesso';
         }
         if(isset($_FILES['printzap']) && !empty($_FILES['printzap']['tmp_name'])){
-            unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/zap/'.$client->printzap);
+            if(!empty($client->printzap)){
+                unlink('C:/xampp/htdocs/goldbanks/works/public/assets/images/media/anexos/zap/'.$client->printzap);
+            }
             $newPrintzap = $_FILES['printzap'];
             $permitidos = ['text/plain'];
             if(in_array($_FILES['printzap']['type'], $permitidos)){
             $zapName = ClientHandler::setTxt($newPrintzap, 'C:\xampp\htdocs\goldbanks\works\public\assets\images\media\anexos\zap');
             }
             $zap = $zapName;
+            
             ClientHandler::editZap($zap, $id);
             $_SESSION['flash'] = 'Conversa zap atualizada com sucesso';
         }
-        if(!empty($name)){
+        if(!empty($name) && $name !== $client->name){
             ClientHandler::editName($name,$id);
             $_SESSION['flash'] = 'Nome do cliente atualizado com sucesso';
         $this->redirect('/viewclient', [
@@ -206,10 +236,9 @@ class ClientController extends Controller {
             'client' => $clients,     
         ]);
         }
-        $_SESSION['flash'] = 'Cliente atualizado com sucesso';
         $this->redirect('/viewclient', [
             'loggedUser' => $this->loggedUser,
-            'client' => $clients,      
+            'client' => $clients,     
         ]);
     }
     public function delClient($id){
